pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "markland"
version = "0.1.0"
description = "MCP server for publishing and sharing markdown documents"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.4.0",
    "python-dotenv>=1.0.0",
    "markdown-it-py>=3.0.0",
    "mdit-py-plugins>=0.4.0",
    "pygments>=2.17.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "jinja2>=3.1.0",
    "sentry-sdk>=2.15.0",
    "resend>=2.5.0",
    "argon2-cffi>=23.1.0",
    "itsdangerous>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.21",
    "httpx>=0.25.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/markland"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
asyncio_default_fixture_loop_scope = "function"
